+++ src/helpers/index.js
@@ -5,7 +5,7 @@
  * @memberof Context
  * @public
  */
-module.exports.filterNotification = require("./filter-notification");
+// module.exports.filterNotification = require("./filter-notification");
 module.exports.requestExtract = require("./request-extract");
 module.exports.handleExtract = require("./handle-extract");
 module.exports.updateSettings = require("./update-settings");