+++ src/middlewares/context-base.js
@@ -6,21 +6,20 @@ import type { HullContextBase, HullRequestBase } from "../types";
  * This middleware is responsible for setting HullContextBase - the base part of the context.
  */
 function contextBaseMiddlewareFactory({
-  instrumentation, queue, cache, connectorConfig
+  instrumentation, queue, cache, connectorConfig, clientConfig
 }: Object) {
   return function contextBaseMiddleware(req: HullRequestBase, res: $Response, next: NextFunction) {
-    const context: HullContextBase = {
-      hostname: req.hostname || "",
-      options: Object.assign({}, req.body, req.query), // body + query
-      connectorConfig,
+    const context = {};
+    context.hostname = req.hostname || "";
+    context.options = Object.assign({}, req.body, req.query); // body + quer;
+    context.clientConfig = clientConfig;
+    context.connectorConfig = connectorConfig;
+    context.service = {};
+    context.cache = cache.getConnectorCache(context);
+    context.metric = instrumentation.getMetric(context);
+    context.enqueue = queue.getEnqueue(context);
 
-      cache: cache.getConnectorCache(),
-      metric: instrumentation.getMetric(),
-      enqueue: queue.getEnqueue(),
-
-      service: {}
-    };
-    req.hull = context;
+    req.hull = (context: HullContextBase);
     next();
   };
 }