+++ src/middlewares/full-context-body.js
@@ -2,6 +2,7 @@
 import type { $Response, NextFunction } from "express";
 import type { HullRequestWithClient } from "../types";
 
+const debug = require("debug")("hull-connector:full-context-fetch-middleware");
 const bodyParser = require("body-parser");
 
 /**
@@ -26,11 +27,29 @@ function bodyFullContextMiddlewareFactory({ requestName }: Object) {
         const timestamp = Math.floor(new Date().getTime() / 1000);
         req.hull.requestId = [requestName, timestamp, body.notification_id].join(":");
       }
+
+      debug("read from body", { connector, users_segments, accounts_segments });
+
+      if (typeof connector !== "object") {
+        next(new Error("Body is missing connector object"));
+      }
+
+      if (!Array.isArray(users_segments)) {
+        next(new Error("Body is missing users_segments array"));
+      }
+
+      if (!Array.isArray(accounts_segments)) {
+        next(new Error("Body is missing accounts_segments array"));
+      }
+
       // $FlowFixMe
-      req.hull = Object.assign({}, req.hull, {
+      req.hull = Object.assign(req.hull, {
+        // $FlowFixMe
         connector,
-        users_segments,
-        accounts_segments,
+        // $FlowFixMe
+        usersSegments: users_segments,
+        // $FlowFixMe
+        accountsSegments: accounts_segments,
         notification: body
       });
       return next();